requirements.txt
streamlit
pandas
reportlab
google-api-python-client
google-auth
